third_party/blink/web_tests/platform/linux/virtual/url-non-special/fast/domurl/url-hash-expected.txt

This is a testharness.js-based test.
[FAIL] non hierarchical
  assert_equals: expected "" but got "#a"
Harness: the test ran to completion.
